pact4s-compiler
Restricted UDT gen sites to templates and blocks (ie, "things which create a scope"). Packages are intentionally excluded in order to force UDTs to always be inner classes. This prevents them from polluting top-level namespaces and ensures compatibility with the existing maven approach for creating jars for individual example programs within a project.