pact4s
Reworked implicit strategy for key selectors and UDFs to get a usable call site of the form unanalyzedXXX(<user function>)

The new strategy uses an implicit conversion to a subclass of scala.reflect.Code[T], which is magic in that the type checker treats instances of Code[FunctionX[…]] as if they were instances of FunctionX[…]. This allows lambdas to be passed to methods which expect a Code[T] without having to explicitly specify the parameter type(s).

Code[T] is actually the marker type used to kick off the Lift phase of the compiler. When a Code[T] is requested, the call site is wrapped in a call to Code.lift[T] as if an implicit conversion were being performed. Because code lifting in Scala is extremely experimental and pretty much broken, we don't actually want to do that. So before we get to that phase, we elide calls to Code.lift and rewrite call sites unanalyzedXXXCode(fun: Code[FunctionY[…]]) to unanalyzedXXX(fun: FunctionY[…]).

Also implemented conversion from unanalyzedUDF(fun) to AnalyzedUDF.default(fun) and removed the default UDF declarations from the example programs.